Added SDL switch and Dreamcast DLC, VMU Sync changes

+++ NullDC CvS2 BEAR/frmDLC.vb
@@ -5,10 +5,11 @@ Imports System.Threading
 Public Class frmDLC
 
     Dim DownloadClient As New WebClient
-    Dim DownloadingGameName = ""
-    Dim DownloadingZipName = ""
+    Dim DownloadingGameName As String = ""
+    Dim DownloadingZipName As String = ""
     Dim DownloadCanceled As Boolean = False
     Dim SelectedGameURL As DownloadableGame = Nothing
+    Dim CurrentDownloadingPlatform As String = ""
 
     Private Sub frmDLC_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         Me.Icon = My.Resources.NewNullDCBearIcon
@@ -28,7 +29,7 @@ Public Class frmDLC
     End Sub
 
     Private Sub ArchiveDotOrgParse(ByVal URL As String, ListView As ListView)
-        Dim StrippedURL = URL.Replace("&output=json", "/")
+        Dim StrippedURL = URL.Replace("&output=json", "")
         StrippedURL = StrippedURL.Replace("/details/", "/download/")
 
         Dim GamesListClient As New WebClient
@@ -47,10 +48,16 @@ Public Class frmDLC
 
                 For Each _sp As String In e.Result.Split("""")
                     If _sp.Contains(".zip") Then
-                        Dim _tmp = _sp.Replace("\/", "").Replace(".zip", "").Replace("_", " ").Trim
-                        Dim _it As New ListViewItem(_tmp)
-                        _it.SubItems.Add(StrippedURL & _sp.Replace("\/", ""))
-                        ListView.Items.Add(_it)
+                        Dim _gameName = _sp.Split("/")(_sp.Split("/").Count - 1).Replace(".zip", "").Replace("_", " ").Trim
+                        Dim _GameLink = StrippedURL & _sp.Replace("\/", "/").Replace("#", "%23")
+
+                        If Not _gameName.Contains("(Windows CE)") Then
+                            Dim _it As New ListViewItem(_gameName)
+                            _it.SubItems.Add(_GameLink)
+                            'Console.WriteLine("Added Game: " & _gameName & " URL: " & _GameLink)
+                            ListView.Items.Add(_it)
+                        End If
+
                     End If
                 Next
 
@@ -66,6 +73,7 @@ Public Class frmDLC
         Try
             ArchiveDotOrgParse("https://archive.org/details/NaomiRomsReuploadByGhostware&output=json", lvGamesList_naomi)
             ArchiveDotOrgParse("https://archive.org/details/AtomiswaveReuploadByGhostware&output=json", lvGamelist_Atomiswave)
+            ArchiveDotOrgParse("https://archive.org/details/DreamcastSelfBoot&output=json", lvGamesList_Dreamcast)
             AddDirectLink("https://archive.org/download/neo-geo-battle-coliseum-unlocked/NeoGeo%20Battle%20Coliseum%20-%20Unlocked.zip", "Neo Geo Battle Coliseum Unlocked", lvGamelist_Atomiswave)
             AddDirectLink("https://archive.org/download/capcom-vs-snk-millenium-fight-2000-unlocked_202010/Capcom_VS_SNK_Millenium_Fight_2000_Unlocked.zip", "Capcom VS SNK Millenium Fight 2000 Unlocked", lvGamesList_naomi)
         Catch ex As Exception
@@ -89,10 +97,18 @@ Public Class frmDLC
         DownloadingGameName = SelectedGameURL.Name
         Dim _tmp As String() = SelectedGameURL.URL.Split("/")
         DownloadingZipName = _tmp(_tmp.Count - 1).Replace(".zip", ".honey")
+        CurrentDownloadingPlatform = SelectedGameURL.Platform
 
         Try
             DownloadClient.Credentials = New NetworkCredential()
-            DownloadClient.DownloadFileTaskAsync(SelectedGameURL.URL, MainformRef.NullDCPath & "\roms\" & DownloadingZipName)
+            Select Case CurrentDownloadingPlatform
+                Case "NA"
+                    DownloadClient.DownloadFileTaskAsync(SelectedGameURL.URL, MainformRef.NullDCPath & "\roms\" & DownloadingZipName)
+
+                Case "DC"
+                    DownloadClient.DownloadFileTaskAsync(SelectedGameURL.URL, MainformRef.NullDCPath & "\dc\roms\" & DownloadingZipName)
+
+            End Select
 
             ProgressBar1.Visible = True
             btnDownload.Text = "Downloading... " & DownloadingGameName
@@ -113,13 +129,28 @@ Public Class frmDLC
         ProgressBar1.Maximum = e.TotalBytesToReceive
         ProgressBar1.Value = e.BytesReceived
         ProgressBar1.Visible = True
-        btnDownload.Text = String.Format("Downloading... {0} (Click again to Cancel)" & vbNewLine & "({1}mb/{2}mb)", DownloadingGameName, Math.Round(e.BytesReceived / 1000000, 2), Math.Round(e.TotalBytesToReceive / 1000000, 2))
+        Dim _NameLength = 20
+        If DownloadingGameName.Length < 20 Then
+            _NameLength = DownloadingGameName.Length
+        End If
+        btnDownload.Text = String.Format("Downloading... {0} (Click again to Cancel)" & vbNewLine & "({1}mb/{2}mb)", DownloadingGameName.Substring(0, _NameLength), Math.Round(e.BytesReceived / 1000000, 2), Math.Round(e.TotalBytesToReceive / 1000000, 2))
     End Sub
 
     Private Sub DownloadComplete()
         Console.WriteLine("Download Done")
-        Dim zipPath = MainformRef.NullDCPath & "\roms\" & DownloadingZipName
-        Dim RomDirectory = MainformRef.NullDCPath & "\roms\" & DownloadingGameName
+
+        Dim zipPath = ""
+        Dim RomDirectory = ""
+
+        Select Case CurrentDownloadingPlatform
+            Case "NA"
+                zipPath = MainformRef.NullDCPath & "\roms\" & DownloadingZipName
+                RomDirectory = MainformRef.NullDCPath & "\roms\" & DownloadingGameName
+            Case "DC"
+                zipPath = MainformRef.NullDCPath & "\dc\roms\" & DownloadingZipName
+                RomDirectory = MainformRef.NullDCPath & "\dc\roms"
+        End Select
+
         'btnDownload.Text = "Installing..."
 
         Dim InstallThread As Thread =
@@ -128,11 +159,14 @@ Public Class frmDLC
                                Try
                                    MainformRef.Invoke(Sub() btnDownload.Text = "Installing...")
 
-                                   If Not Directory.Exists(RomDirectory) Then
-                                       Directory.CreateDirectory(RomDirectory)
-                                   Else
-                                       Directory.Delete(RomDirectory, True)
-                                       Directory.CreateDirectory(RomDirectory)
+                                   If CurrentDownloadingPlatform = "NA" Then ' Naomi has to delete the whole folder if it exists
+                                       If Not Directory.Exists(RomDirectory) Then
+                                           Directory.CreateDirectory(RomDirectory)
+                                       Else
+                                           Directory.Delete(RomDirectory, True)
+                                           Directory.CreateDirectory(RomDirectory)
+                                       End If
+
                                    End If
 
                                    If File.Exists(zipPath) Then
@@ -175,7 +209,12 @@ Public Class frmDLC
     End Sub
 
     Private Sub lnkRoms_LinkClicked(sender As Object, e As LinkLabelLinkClickedEventArgs) Handles lnkRoms.LinkClicked
-        Process.Start("https://archive.org/details/NaomiRomsReuploadByGhostware")
+        If tc_games.SelectedIndex = 0 Or tc_games.SelectedIndex = 1 Then
+            Process.Start("https://archive.org/details/NaomiRomsReuploadByGhostware")
+        Else
+            Process.Start("https://archive.org/details/DreamcastSelfBoot")
+        End If
+
     End Sub
 
     Private Sub btnRomsFolder_Click(sender As Object, e As EventArgs) Handles btnRomsFolder.Click
@@ -198,7 +237,7 @@ Public Class frmDLC
         If lvGamelist_Atomiswave.SelectedItems.Count = 0 Then
             SelectedGameURL = Nothing
         Else
-            SelectedGameURL = New DownloadableGame(lvGamelist_Atomiswave.SelectedItems(0).SubItems(1).Text, lvGamelist_Atomiswave.SelectedItems(0).SubItems(0).Text)
+            SelectedGameURL = New DownloadableGame(lvGamelist_Atomiswave.SelectedItems(0).SubItems(1).Text, lvGamelist_Atomiswave.SelectedItems(0).SubItems(0).Text, "NA")
         End If
 
     End Sub
@@ -207,7 +246,16 @@ Public Class frmDLC
         If lvGamesList_naomi.SelectedItems.Count = 0 Then
             SelectedGameURL = Nothing
         Else
-            SelectedGameURL = New DownloadableGame(lvGamesList_naomi.SelectedItems(0).SubItems(1).Text, lvGamesList_naomi.SelectedItems(0).SubItems(0).Text)
+            SelectedGameURL = New DownloadableGame(lvGamesList_naomi.SelectedItems(0).SubItems(1).Text, lvGamesList_naomi.SelectedItems(0).SubItems(0).Text, "NA")
+        End If
+
+    End Sub
+
+    Private Sub lvGamesList_Dreamcast_SelectedIndexChanged(sender As Object, e As EventArgs) Handles lvGamesList_Dreamcast.SelectedIndexChanged
+        If lvGamesList_Dreamcast.SelectedItems.Count = 0 Then
+            SelectedGameURL = Nothing
+        Else
+            SelectedGameURL = New DownloadableGame(lvGamesList_Dreamcast.SelectedItems(0).SubItems(1).Text, lvGamesList_Dreamcast.SelectedItems(0).SubItems(0).Text, "DC")
         End If
 
     End Sub
@@ -223,7 +271,7 @@ Public Class frmDLC
         End Select
     End Sub
 
-    Private Sub LinkLabel1_LinkClicked(sender As Object, e As LinkLabelLinkClickedEventArgs) Handles LinkLabel1.LinkClicked
+    Private Sub LinkLabel1_LinkClicked(sender As Object, e As LinkLabelLinkClickedEventArgs)
         Process.Start("https://cdromance.com/dc-iso/?sorted=downloads")
     End Sub
 End Class
@@ -232,10 +280,13 @@ Class DownloadableGame
 
     Public URL = ""
     Public Name = ""
+    Public Platform = ""
 
-    Public Sub New(ByVal _url As String, ByVal _name As String)
+    Public Sub New(ByVal _url As String, ByVal _name As String, ByVal _platform As String)
         URL = _url
         Name = _name
+        Platform = _platform
+
     End Sub
 
 End Class
